requirements.txt
# ProjectX Trading Bot Requirements

# ProjectX SDK (Official)
project-x-py>=0.1.0

# Data Processing
pandas>=2.1.0
polars>=0.19.0
numpy>=1.24.0

# Technical Analysis
ta-lib>=0.4.0

# API & Networking
httpx>=0.25.0
websockets>=12.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Visualization (for backtesting reports)
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
rich>=13.7.0  # For beautiful console output

# Optimization
optuna>=3.5.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
